Refactor parsing tests
